incusd/storage: Add support for creating, deleting, restoring qcow2 snapshots

+++ internal/server/storage/backend.go
@@ -54,6 +54,7 @@ import (
 	"github.com/lxc/incus/v6/shared/ioprogress"
 	"github.com/lxc/incus/v6/shared/logger"
 	"github.com/lxc/incus/v6/shared/revert"
+	"github.com/lxc/incus/v6/shared/subprocess"
 	"github.com/lxc/incus/v6/shared/units"
 	"github.com/lxc/incus/v6/shared/util"
 )
@@ -3084,9 +3085,17 @@ func (b *backend) CreateInstanceSnapshot(inst instance.Instance, src instance.In
 
 	defer unlock()
 
-	err = b.driver.CreateVolumeSnapshot(vol, op)
-	if err != nil {
-		return err
+	if srcDBVol.Config["block.type"] == "qcow2" {
+		_, snapName, _ := api.GetParentAndSnapshotName(vol.Name())
+		err = src.CreateInternalSnapshot(snapName)
+		if err != nil {
+			return err
+		}
+	} else {
+		err = b.driver.CreateVolumeSnapshot(vol, op)
+		if err != nil {
+			return err
+		}
 	}
 
 	err = b.ensureInstanceSnapshotSymlink(inst.Type(), inst.Project().Name, inst.Name())
@@ -3322,40 +3331,47 @@ func (b *backend) RestoreInstanceSnapshot(inst instance.Instance, src instance.I
 		})
 	}
 
-	err = b.driver.RestoreVolume(vol, snapshotName, op)
-	if err != nil {
-		var snapErr drivers.ErrDeleteSnapshots
-		if errors.As(err, &snapErr) {
-			// We need to delete some snapshots and try again.
-			snaps, err := inst.Snapshots()
-			if err != nil {
-				return err
-			}
-
-			// Go through all the snapshots.
-			for _, snap := range snaps {
-				_, snapName, _ := api.GetParentAndSnapshotName(snap.Name())
-				if !slices.Contains(snapErr.Snapshots, snapName) {
-					continue
-				}
-
-				// Delete snapshot instance if listed in the error as one that needs removing.
-				err := snap.Delete(true)
+	if srcDBVol.Config["block.type"] == "qcow2" {
+		err = b.restoreInternalVolume(vol, snapshotName, op)
+		if err != nil {
+			return err
+		}
+	} else {
+		err = b.driver.RestoreVolume(vol, snapshotName, op)
+		if err != nil {
+			var snapErr drivers.ErrDeleteSnapshots
+			if errors.As(err, &snapErr) {
+				// We need to delete some snapshots and try again.
+				snaps, err := inst.Snapshots()
 				if err != nil {
 					return err
 				}
+
+				// Go through all the snapshots.
+				for _, snap := range snaps {
+					_, snapName, _ := api.GetParentAndSnapshotName(snap.Name())
+					if !slices.Contains(snapErr.Snapshots, snapName) {
+						continue
+					}
+
+					// Delete snapshot instance if listed in the error as one that needs removing.
+					err := snap.Delete(true)
+					if err != nil {
+						return err
+					}
+				}
+
+				// Now try restoring again.
+				err = b.driver.RestoreVolume(vol, snapshotName, op)
+				if err != nil {
+					return err
+				}
+
+				return nil
 			}
 
-			// Now try restoring again.
-			err = b.driver.RestoreVolume(vol, snapshotName, op)
-			if err != nil {
-				return err
-			}
-
-			return nil
+			return err
 		}
-
-		return err
 	}
 
 	reverter.Success()
@@ -6035,10 +6051,17 @@ func (b *backend) CreateCustomVolumeSnapshot(projectName, volName string, newSna
 
 	defer unlock()
 
-	// Create the snapshot on the storage device.
-	err = b.driver.CreateVolumeSnapshot(vol, op)
-	if err != nil {
-		return err
+	if parentVol.Config["block.type"] == "qcow2" {
+		err = b.createInternalVolumeSnapshot(vol, op)
+		if err != nil {
+			return err
+		}
+	} else {
+		// Create the snapshot on the storage device.
+		err = b.driver.CreateVolumeSnapshot(vol, op)
+		if err != nil {
+			return err
+		}
 	}
 
 	b.state.Events.SendLifecycle(projectName, lifecycle.StorageVolumeSnapshotCreated.Event(vol, string(vol.Type()), projectName, op, logger.Ctx{"type": vol.Type()}))
@@ -6144,7 +6167,12 @@ func (b *backend) DeleteCustomVolumeSnapshot(projectName, volName string, op *op
 		return err
 	}
 
-	if volExists {
+	if volume.Config["block.type"] == "qcow2" {
+		err = b.deleteInternalVolumeSnapshot(vol, op)
+		if err != nil {
+			return err
+		}
+	} else if volExists {
 		err := b.driver.DeleteVolumeSnapshot(vol, op)
 		if err != nil {
 			return err
@@ -6214,26 +6242,33 @@ func (b *backend) RestoreCustomVolume(projectName, volName string, snapshotName 
 	volStorageName := project.StorageVolume(projectName, volName)
 	vol := b.GetVolume(drivers.VolumeTypeCustom, contentType, volStorageName, curVol.Config)
 
-	err = b.driver.RestoreVolume(vol, snapshotName, op)
-	if err != nil {
-		var snapErr drivers.ErrDeleteSnapshots
-		if errors.As(err, &snapErr) {
-			// We need to delete some snapshots and try again.
-			for _, snapName := range snapErr.Snapshots {
-				err := b.DeleteCustomVolumeSnapshot(projectName, fmt.Sprintf("%s/%s", volName, snapName), op)
+	if curVol.Config["block.type"] == "qcow2" {
+		err = b.restoreInternalVolume(vol, snapshotName, op)
+		if err != nil {
+			return err
+		}
+	} else {
+		err = b.driver.RestoreVolume(vol, snapshotName, op)
+		if err != nil {
+			var snapErr drivers.ErrDeleteSnapshots
+			if errors.As(err, &snapErr) {
+				// We need to delete some snapshots and try again.
+				for _, snapName := range snapErr.Snapshots {
+					err := b.DeleteCustomVolumeSnapshot(projectName, fmt.Sprintf("%s/%s", volName, snapName), op)
+					if err != nil {
+						return err
+					}
+				}
+
+				// Now try again.
+				err = b.driver.RestoreVolume(vol, snapshotName, op)
 				if err != nil {
 					return err
 				}
 			}
 
-			// Now try again.
-			err = b.driver.RestoreVolume(vol, snapshotName, op)
-			if err != nil {
-				return err
-			}
+			return err
 		}
-
-		return err
 	}
 
 	b.state.Events.SendLifecycle(projectName, lifecycle.StorageVolumeRestored.Event(vol, string(vol.Type()), projectName, op, logger.Ctx{"snapshot": snapshotName}))
@@ -7539,4 +7574,76 @@ func (b *backend) getFirstAdminStorageBucketPoolKey(projectName string, bucketNa
 	}
 
 	return bucketKey, nil
+}
+
+func (b *backend) createInternalVolumeSnapshot(snapVol drivers.Volume, op *operations.Operation) error {
+	parentName, snapName, _ := api.GetParentAndSnapshotName(snapVol.Name())
+	parentVol := b.GetVolume(snapVol.Type(), snapVol.ContentType(), parentName, snapVol.Config())
+
+	err := parentVol.MountTask(func(mountPath string, op *operations.Operation) error {
+		// Get the device path.
+		devPath, err := b.driver.GetVolumeDiskPath(parentVol)
+		if err != nil {
+			return err
+		}
+
+		_, err = subprocess.RunCommand("qemu-img", "snapshot", "-c", snapName, devPath)
+		if err != nil {
+			return err
+		}
+
+		return nil
+	}, op)
+	if err != nil {
+		return fmt.Errorf("Error creating qcow2 volume snapshot: %w", err)
+	}
+
+	return nil
+}
+
+func (b *backend) deleteInternalVolumeSnapshot(snapVol drivers.Volume, op *operations.Operation) error {
+	parentName, snapName, _ := api.GetParentAndSnapshotName(snapVol.Name())
+	parentVol := b.GetVolume(snapVol.Type(), snapVol.ContentType(), parentName, snapVol.Config())
+
+	err := parentVol.MountTask(func(mountPath string, op *operations.Operation) error {
+		// Get the device path.
+		devPath, err := b.driver.GetVolumeDiskPath(parentVol)
+		if err != nil {
+			return err
+		}
+
+		_, err = subprocess.RunCommand("qemu-img", "snapshot", "-d", snapName, devPath)
+		if err != nil {
+			return err
+		}
+
+		return nil
+	}, op)
+	if err != nil {
+		return fmt.Errorf("Error deleting qcow2 volume snapshot: %w", err)
+	}
+
+	return nil
+}
+
+func (b *backend) restoreInternalVolume(vol drivers.Volume, snapshotName string, op *operations.Operation) error {
+	err := vol.MountTask(func(mountPath string, op *operations.Operation) error {
+		// Get the device path.
+		devPath, err := b.driver.GetVolumeDiskPath(vol)
+		if err != nil {
+			return err
+		}
+
+		_, err = subprocess.RunCommand("qemu-img", "snapshot", "-a", snapshotName, devPath)
+		if err != nil {
+			return err
+		}
+
+		return nil
+	}, op)
+	if err != nil {
+		return fmt.Errorf("Error restoring qcow2 volume snapshot: %w", err)
+	}
+
+	return nil
 }
